# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Easylife API
  version: 1.1.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Easylife 前后端 V1 接口契约。

    当前版本只冻结情绪分析接口。该接口是无状态 AI 推理接口，不负责创建、
    更新或保存情绪日记。只有用户明确确认记录后，客户端才应调用未来的数据同步接口。

    V1 兼容规则：
    - 可以新增可选字段。
    - 不得删除字段、改变字段类型或改变既有字段语义。
    - 破坏性变更必须发布新的主版本路径。
servers:
  - url: https://api.example.com
    description: 示例生产环境
  - url: http://localhost:3000
    description: 本地开发环境
...
paths:
  /v1/auth/sms/codes:
    post:
      tags:
        - Authentication
      operationId: sendSmsCode
      summary: 发送短信验证码
      description: |
        向中国大陆手机号发送验证码。验证码 5 分钟有效，同一手机号 60 秒后方可重发。
        同一手机号每小时最多发送 5 次、每天最多 10 次；同一设备或 IP 每小时最多
        20 次。服务端不得通过响应泄露手机号是否已注册。
      security: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendSmsCodeRequest"
            examples:
              login:
                $ref: "#/components/examples/SendSmsCodeRequest"
      responses:
        "200":
          description: 验证码发送请求已接受
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendSmsCodeResponse"
              examples:
                default:
                  $ref: "#/components/examples/SendSmsCodeResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
        "503":
          $ref: "#/components/responses/SmsProviderUnavailableError"
  /v1/auth/sms/verify:
    post:
      tags:
        - Authentication
      operationId: verifySmsCode
      summary: 验证短信验证码
      description: |
        登录验证码验证成功后，如果手机号尚未注册则自动创建账号，并返回登录会话。
        注销验证码验证成功后只返回短期注销凭证，不创建新会话。
        每个 challenge 最多允许 5 次错误验证，超过后立即失效。
      security: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifySmsCodeRequest"
            examples:
              login:
                $ref: "#/components/examples/VerifySmsCodeRequest"
      responses:
        "200":
          description: 验证成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/LoginVerificationResponse"
                  - $ref: "#/components/schemas/AccountDeletionVerificationResponse"
              examples:
                login:
                  $ref: "#/components/examples/LoginVerificationResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "410":
          $ref: "#/components/responses/SmsCodeExpiredError"
        "422":
          $ref: "#/components/responses/SmsCodeInvalidError"
        "429":
          $ref: "#/components/responses/VerificationAttemptsExceededError"
  /v1/auth/token/refresh:
    post:
      tags:
        - Authentication
      operationId: refreshAccessToken
      summary: 刷新访问令牌
      description: 使用一次性轮换的刷新令牌签发新会话；旧刷新令牌成功使用后立即失效。
      security: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: 令牌刷新成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/InvalidRefreshTokenError"
  /v1/auth/logout:
    post:
      tags:
        - Authentication
      operationId: logout
      summary: 退出当前设备
      description: 撤销当前设备会话及对应刷新令牌；该操作必须幂等。
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "204":
          description: 当前设备会话已撤销或此前已经失效
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /v1/emotion/analyze:
    post:
      tags:
        - Emotion
      operationId: analyzeEmotion
      summary: 分析用户当前情绪
      description: |
        根据用户本轮输入和经过最小化处理的个性化上下文，返回结构化情绪洞察。
        服务端不得进行医学诊断，也不得在该接口中自动保存情绪日记。
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmotionAnalyzeRequest"
            examples:
              default:
                $ref: "#/components/examples/EmotionAnalyzeRequest"
      responses:
        "200":
          description: 情绪分析成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmotionAnalyzeResponse"
              examples:
                default:
                  $ref: "#/components/examples/EmotionAnalyzeResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "413":
          $ref: "#/components/responses/PayloadTooLargeError"
        "422":
          $ref: "#/components/responses/AiOutputInvalidError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
        "503":
          $ref: "#/components/responses/AiProviderUnavailableError"
  /v1/companion/reply:
    post:
      tags:
        - Companion
      operationId: createCompanionReply
      summary: 生成陪伴式多轮回复
      description: |
        根据本轮已授权对话和最小化上下文生成一条朋友式陪伴回复。该接口只负责
        水上对话体验，不创建情绪日记，也不保存长期记忆。
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanionReplyRequest"
            examples:
              default:
                $ref: "#/components/examples/CompanionReplyRequest"
      responses:
        "200":
          description: 陪伴回复生成成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanionReplyResponse"
              examples:
                default:
                  $ref: "#/components/examples/CompanionReplyResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "413":
          $ref: "#/components/responses/PayloadTooLargeError"
        "422":
          $ref: "#/components/responses/AiOutputInvalidError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
        "503":
          $ref: "#/components/responses/AiProviderUnavailableError"
  /v1/emotion-journals/summarize:
    post:
      tags:
        - Companion
      operationId: summarizeEmotionJournal
      summary: 整理本轮陪伴对话为情绪日记
      description: |
        仅在用户主动点击保存情绪日记后调用。服务端根据完整本轮对话整理结构化日记
        内容；调用该接口本身不写数据库，保存仍由客户端本地优先与同步流程完成。
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmotionJournalSummaryRequest"
            examples:
              default:
                $ref: "#/components/examples/EmotionJournalSummaryRequest"
      responses:
        "200":
          description: 情绪日记整理成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmotionJournalSummaryResponse"
              examples:
                default:
                  $ref: "#/components/examples/EmotionJournalSummaryResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "413":
          $ref: "#/components/responses/PayloadTooLargeError"
        "422":
          $ref: "#/components/responses/AiOutputInvalidError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
        "503":
          $ref: "#/components/responses/AiProviderUnavailableError"
  /v1/memories/extract:
    post:
      tags:
        - Memory
      operationId: extractLongTermMemoryCandidates
      summary: 从已确认保存内容中提炼长期记忆候选
      description: |
        只从用户已经确认保存的内容中提炼低敏、稳定、可复用的记忆候选。接口不直接
        修改用户记忆列表，最终增删改仍由客户端或同步写入流程确认。
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryExtractRequest"
            examples:
              default:
                $ref: "#/components/examples/MemoryExtractRequest"
      responses:
        "200":
          description: 长期记忆候选提炼成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryExtractResponse"
              examples:
                default:
                  $ref: "#/components/examples/MemoryExtractResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "413":
          $ref: "#/components/responses/PayloadTooLargeError"
        "422":
          $ref: "#/components/responses/AiOutputInvalidError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
        "503":
          $ref: "#/components/responses/AiProviderUnavailableError"
  /v1/pet-avatar/generate:
    post:
      tags:
        - Pet
      operationId: generatePetAvatar
      summary: 生成伙伴形象
      description: |
        在用户明确授权后，根据上传的伙伴照片生成陪伴形象。请求使用 data URL 传输
        PNG、JPEG 或 WebP 图片；服务端调用图像生成 provider，不在 Flutter 客户端
        保存模型密钥。
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetAvatarGenerateRequest"
            examples:
              default:
                $ref: "#/components/examples/PetAvatarGenerateRequest"
      responses:
        "200":
          description: 伙伴形象生成成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetAvatarGenerateResponse"
              examples:
                default:
                  $ref: "#/components/examples/PetAvatarGenerateResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "413":
          $ref: "#/components/responses/PayloadTooLargeError"
        "422":
          $ref: "#/components/responses/AiOutputInvalidError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
        "503":
          $ref: "#/components/responses/AiProviderUnavailableError"
  /v1/sync/push:
    post:
      tags:
        - Sync
      operationId: pushSyncMutations
      summary: 上传本地待同步变更
      description: |
        批量上传本地 Outbox 中的变更。mutationId 在用户范围内必须幂等；
        重复提交同一 mutationId 必须返回之前的处理结果，不得重复创建数据。
        情绪分析结果只有在用户明确确认记录后才能作为 emotionEntry 或 memoryNote 上传。
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncPushRequest"
            examples:
              default:
                $ref: "#/components/examples/SyncPushRequest"
      responses:
        "200":
          description: 批次已处理；单条成功、冲突或拒绝结果在 results 中返回
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncPushResponse"
              examples:
                default:
                  $ref: "#/components/examples/SyncPushResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "413":
          $ref: "#/components/responses/PayloadTooLargeError"
  /v1/sync/pull:
    get:
      tags:
        - Sync
      operationId: pullSyncChanges
      summary: 拉取服务端增量变更
      description: |
        使用不透明游标按稳定顺序拉取增量数据，包括删除墓碑。首次同步不传 cursor。
        客户端必须持续拉取直到 hasMore 为 false，再保存最终 nextCursor。
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: cursor
          in: query
          required: false
          description: 服务端返回的不透明同步游标。
          schema:
            type: string
            maxLength: 500
        - name: limit
          in: query
          required: false
          description: 单页最大变更数量。
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
      responses:
        "200":
          description: 增量变更列表
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncPullResponse"
              examples:
                default:
                  $ref: "#/components/examples/SyncPullResponse"
        "400":
          $ref: "#/components/responses/InvalidCursorError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /v1/me/account:
    delete:
      tags:
        - Account
      operationId: deleteAccount
      summary: 注销当前账号
      description: |
        使用短信验证后签发的短期注销凭证启动账号注销。成功后立即撤销全部会话，
        客户端必须停止同步并清除本机令牌和业务数据。服务端异步完成云端物理删除。
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountRequest"
            examples:
              default:
                $ref: "#/components/examples/DeleteAccountRequest"
      responses:
        "202":
          description: 注销任务已创建，全部会话已撤销
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAccountResponse"
              examples:
                default:
                  $ref: "#/components/examples/DeleteAccountResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "410":
          $ref: "#/components/responses/DeletionVerificationExpiredError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 未来真实登录签发的访问令牌。
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: true
      description: 客户端生成的 UUID，用于端到端请求追踪。
      schema:
        type: string
        format: uuid
      example: 6f2aa37d-e95d-4b52-8df4-0cf3e17e5188
  headers:
    RequestId:
      required: true
      description: 与请求一致的追踪 ID；若请求 ID 无效，服务端可生成新 UUID。
      schema:
        type: string
        format: uuid
      example: 6f2aa37d-e95d-4b52-8df4-0cf3e17e5188
  schemas:
...
    SendSmsCodeRequest:
      type: object
      additionalProperties: false
      required:
        - phone
        - purpose
        - deviceId
      properties:
        phone:
          $ref: "#/components/schemas/ChinaMainlandPhone"
        purpose:
          $ref: "#/components/schemas/SmsPurpose"
        deviceId:
          $ref: "#/components/schemas/DeviceId"
    SendSmsCodeResponse:
      type: object
      additionalProperties: false
      required:
        - challengeId
        - expiresIn
        - resendAfter
      properties:
        challengeId:
          type: string
          format: uuid
          description: 本次验证码挑战 ID，不包含验证码本身。
        expiresIn:
          type: integer
          const: 300
          description: 验证码有效秒数。
        resendAfter:
          type: integer
          const: 60
          description: 可以再次请求验证码前的等待秒数。
    VerifySmsCodeRequest:
      type: object
      additionalProperties: false
      required:
        - challengeId
        - phone
        - code
        - deviceId
      properties:
        challengeId:
          type: string
          format: uuid
        phone:
          $ref: "#/components/schemas/ChinaMainlandPhone"
        code:
          type: string
          pattern: '^[0-9]{6}$'
          description: 6 位数字短信验证码。
          example: "123456"
        deviceId:
          $ref: "#/components/schemas/DeviceId"
...
    LoginVerificationResponse:
      type: object
      additionalProperties: false
      required:
        - purpose
        - isNewUser
        - user
        - tokens
      properties:
        purpose:
          type: string
          const: login
        isNewUser:
          type: boolean
          description: true 表示手机号首次验证成功且账号已自动注册。
        user:
          $ref: "#/components/schemas/AuthUser"
        tokens:
          $ref: "#/components/schemas/TokenPair"
    AccountDeletionVerificationResponse:
      type: object
      additionalProperties: false
      required:
        - purpose
        - deletionToken
        - expiresIn
      properties:
        purpose:
          type: string
          const: account_deletion
        deletionToken:
          type: string
          minLength: 20
          description: 仅可用于注销当前账号的一次性短期凭证。
        expiresIn:
          type: integer
          const: 600
          description: 注销凭证有效期 10 分钟。
...
  responses:
    ValidationError:
      description: 请求字段缺失、格式错误或超出字段限制
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: 请求内容不符合接口要求
              requestId: 6f2aa37d-e95d-4b52-8df4-0cf3e17e5188
...
    SmsCodeExpiredError:
      description: 验证码已过期、已使用或 challenge 已失效
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: SMS_CODE_EXPIRED
              message: 验证码已失效，请重新获取
              requestId: 6f2aa37d-e95d-4b52-8df4-0cf3e17e5188
    SmsCodeInvalidError:
      description: 验证码错误但尚未达到错误次数上限
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: SMS_CODE_INVALID
              message: 验证码错误
              requestId: 6f2aa37d-e95d-4b52-8df4-0cf3e17e5188
...
    SmsProviderUnavailableError:
      description: 短信服务暂时不可用
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
        Retry-After:
          description: 建议客户端等待的秒数。
          schema:
            type: integer
            minimum: 1
          example: 60
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: SMS_PROVIDER_UNAVAILABLE
              message: 短信服务暂时不可用
              requestId: 6f2aa37d-e95d-4b52-8df4-0cf3e17e5188
...
    RateLimitedError:
      description: 请求频率超过服务端限制
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
        Retry-After:
          description: 建议客户端等待的秒数。
          schema:
            type: integer
            minimum: 1
          example: 30
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: RATE_LIMITED
              message: 请求过于频繁，请稍后再试
              requestId: 6f2aa37d-e95d-4b52-8df4-0cf3e17e5188
...
  examples:
    SendSmsCodeRequest:
      summary: 请求登录验证码
      value:
        phone: "+8613812345678"
        purpose: login
        deviceId: 6ecb2ba5-6c51-4a40-b908-8be4311c7f85
    SendSmsCodeResponse:
      summary: 验证码挑战
      value:
        challengeId: 984e6346-f8a1-4511-8ec2-a960bc338705
        expiresIn: 300
        resendAfter: 60
    VerifySmsCodeRequest:
      summary: 验证登录验证码
      value:
        challengeId: 984e6346-f8a1-4511-8ec2-a960bc338705
        phone: "+8613812345678"
        code: "123456"
        deviceId: 6ecb2ba5-6c51-4a40-b908-8be4311c7f85
    LoginVerificationResponse:
      summary: 新手机号自动注册并登录
      value:
        purpose: login
        isNewUser: true
        user:
          id: e0e00cc4-f640-4975-b21f-ad93c28b67f0
          phoneMasked: "138****5678"
        tokens:
          accessToken: example-access-token-with-at-least-twenty-characters
          accessTokenExpiresIn: 900
          refreshToken: example-refresh-token-with-at-least-twenty-characters
          refreshTokenExpiresIn: 2592000
...